# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample document for API.",
        "title": "Neyu Document API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Son Nguyen",
            "email": "son.nguyen.2@neyu.co"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:9091",
    "basePath": "/api",
    "paths": {
        "/user": {
            "get": {
                "description": "Get list user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get user data",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.User"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Insert user data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Insert user data",
                "parameters": [
                    {
                        "description": "User body",
                        "name": "offset",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    }
                }
            }
        },
        "/user/batch": {
            "post": {
                "description": "Insert batch data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Insert batch data",
                "parameters": [
                    {
                        "description": "list user body",
                        "name": "offset",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.User"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.User": {
            "type": "object",
            "required": [
                "jphone",
                "juser_name"
            ],
            "properties": {
                "jphone": {
                    "type": "string"
                },
                "juser_id": {
                    "type": "integer"
                },
                "juser_name": {
                    "type": "string"
                }
            }
        }
    }
}